change to using fat-arrow for __express method
Previously, apps were crashing with 500 TypeError: Object #<View> has no method '_readFile'. Artifact of refactor, where __express function was moved to a method on the 'HulkHogan' class. Since application authors hand over __express method to the framework, the 'this' value was being lost. Changing to a fat-arrow binds this to the HulkHogan instance, and a reference to @_readfile is restored.